backend/requirements.txt
flask
flask-cors
flask-restful
psycopg2-binary
pytest
pytest-flask
scikit-learn
pandas
tensorflow
